+++ NEWS.md
@@ -3,6 +3,10 @@ Phan NEWS
 ?? ??? 2019, Phan 2.0.0 (dev)
 -----------------------
 
+New features(CLI, Configs):
++ Enable language server features by default. (#2358)
+  New CLI flags to disable features: `--language-server-disable-hover`, `--language-server-disable-go-to-definition`, `--language-server-disable-completion`
+
 Backwards Incompatible Changes:
 + Require php-ast 1.0.1 or newer (or the absence of php-ast with `--allow-polyfill-parser`)
   Phan switched from using [AST version 50 to version 70](https://github.com/nikic/php-ast#ast-versioning).